+++ tests/integration/models/cpack.py
@@ -97,9 +97,10 @@ m = Model(
 
 # solve with a good global solver
 print("Starting solve, be patient (log only shown afterwards)...")
-m.solve(options=Options(relative_optimality_gap=0.01))
+m.solve(solver="scip", options=Options(relative_optimality_gap=0.01))
 
-assert math.isclose(m.objective_value, 0.3702, rel_tol=0.001)
+print(f"{m.objective_value=}")
+assert math.isclose(m.objective_value, 0.3701919131257)
 
 rval = r.records.loc[0, "level"]
 print("Maximized radius:", rval)